requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.6
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.29.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzlocal==5.0.1
urllib3==1.26.15
